Update ScrnShotgun.uc
tripwire didn't update the reloadanimrate when they tweaked the shotgun reloadrate. Currently the animation plays too fast and if you do a 8 shell reload it will also play a bit of an unused portion of the shotgun reload (a pump animation). Since that part of the reload animation isn't being used for anything yet, the animation should be slowed down to match the reload time. According to the SDK there are 18 frames between each shell load, so (18/30)/0.6667 gives a new reloadanimrate of 0.90 for perfect sync